refactor search drawer for access with tab and escape keys
